requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.20
psycopg-binary==3.1.18
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn
python-dateutil